add proposal interfaces
