+++ packages/app/interfaces/proposals.tsx
@@ -1,0 +1,34 @@
+import { BigNumber } from 'ethers';
+
+export enum Status {
+  Open,
+  Challenge,
+  Completed,
+  Passed,
+  Failed,
+  All,
+}
+
+export interface BaseProposal {
+  name: string;
+  missionStatement: string;
+  ethereumAddress: string;
+  profileImage: string;
+  votesFor: BigNumber;
+  votesAgainst: BigNumber;
+  status: Status;
+  stageDeadline: Date;
+}
+
+export interface BeneficiaryProposal extends BaseProposal {
+  headerImage: string;
+  twitterUrl?: string;
+  linkedinUrl?: string;
+  facebookUrl?: string;
+  instagramUrl?: string;
+  githubUrl?: string;
+  impactReports?: string[];
+  additionalImages?: string[];
+  proofOfOwnership: string;
+  id: number;
+}
