+++ packages/app/pages/beneficiary-proposals/[id].tsx
@@ -1,6 +1,6 @@
 import BeneficiaryPage from 'components/Beneficiaries/BeneficiaryPage/BeneficiaryPage';
 import { ContractsContext } from 'context/Web3/contracts';
-import { BeneficiaryProposal } from 'interfaces/beneficiaries';
+import { BeneficiaryProposal } from 'interfaces/proposals';
 import router from 'next/router';
 import { useContext, useEffect, useState } from 'react';
 import { getProposal } from 'utils/getProposals';
